Updates to temporal_capi for the previous changes

+++ temporal_capi/bindings/cpp/temporal_rs/DateDuration.hpp
@@ -20,7 +20,7 @@ namespace capi {
     extern "C" {
     
     typedef struct temporal_rs_DateDuration_new_result {union {temporal_rs::capi::DateDuration* ok; temporal_rs::capi::TemporalError err;}; bool is_ok;} temporal_rs_DateDuration_new_result;
-    temporal_rs_DateDuration_new_result temporal_rs_DateDuration_new(double years, double months, double weeks, double days);
+    temporal_rs_DateDuration_new_result temporal_rs_DateDuration_new(int64_t years, int64_t months, int64_t weeks, int64_t days);
     
     temporal_rs::capi::DateDuration* temporal_rs_DateDuration_abs(const temporal_rs::capi::DateDuration* self);
     
@@ -35,7 +35,7 @@ namespace capi {
 } // namespace capi
 } // namespace
 
-inline diplomat::result<std::unique_ptr<temporal_rs::DateDuration>, temporal_rs::TemporalError> temporal_rs::DateDuration::new_(double years, double months, double weeks, double days) {
+inline diplomat::result<std::unique_ptr<temporal_rs::DateDuration>, temporal_rs::TemporalError> temporal_rs::DateDuration::new_(int64_t years, int64_t months, int64_t weeks, int64_t days) {
   auto result = temporal_rs::capi::temporal_rs_DateDuration_new(years,
     months,
     weeks,
