+++ temporal_capi/bindings/cpp/temporal_rs/Duration.d.hpp
@@ -33,9 +33,9 @@ namespace temporal_rs {
 class Duration {
 public:
 
-  inline static diplomat::result<std::unique_ptr<temporal_rs::Duration>, temporal_rs::TemporalError> create(double years, double months, double weeks, double days, double hours, double minutes, double seconds, double milliseconds, double microseconds, double nanoseconds);
+  inline static diplomat::result<std::unique_ptr<temporal_rs::Duration>, temporal_rs::TemporalError> create(int64_t years, int64_t months, int64_t weeks, int64_t days, int64_t hours, int64_t minutes, int64_t seconds, int64_t milliseconds, int64_t microseconds, int64_t nanoseconds);
 
-  inline static diplomat::result<std::unique_ptr<temporal_rs::Duration>, temporal_rs::TemporalError> from_day_and_time(double day, const temporal_rs::TimeDuration& time);
+  inline static diplomat::result<std::unique_ptr<temporal_rs::Duration>, temporal_rs::TemporalError> from_day_and_time(int64_t day, const temporal_rs::TimeDuration& time);
 
   inline static diplomat::result<std::unique_ptr<temporal_rs::Duration>, temporal_rs::TemporalError> from_partial_duration(temporal_rs::PartialDuration partial);
 
@@ -45,25 +45,25 @@ public:
 
   inline const temporal_rs::DateDuration& date() const;
 
-  inline double years() const;
+  inline int64_t years() const;
 
-  inline double months() const;
+  inline int64_t months() const;
 
-  inline double weeks() const;
+  inline int64_t weeks() const;
 
-  inline double days() const;
+  inline int64_t days() const;
 
-  inline double hours() const;
+  inline int64_t hours() const;
 
-  inline double minutes() const;
+  inline int64_t minutes() const;
 
-  inline double seconds() const;
+  inline int64_t seconds() const;
 
-  inline double milliseconds() const;
+  inline int64_t milliseconds() const;
 
-  inline double microseconds() const;
+  inline int64_t microseconds() const;
 
-  inline double nanoseconds() const;
+  inline int64_t nanoseconds() const;
 
   inline temporal_rs::Sign sign() const;
